core/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap4==2022.1
Django==4.2.5
django-cors-headers==4.2.0
django-crispy-forms==2.0
djangorestframework==3.14.0
html-sanitizer==2.2.0
lxml==4.9.3
Pillow==10.0.0
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
